requirements.txt
  
dlib
imutils
numpy
opencv-python
playsound
scipy
tensorflow-cpu
torch
torchvision
tqdm
pandas
flask
flask-cors
paho-mqtt
requests